Add various FHIR types
New types are: UsageContext, Range, Identifier, Period,
ContactDetail, ContactPoint, Narrative, Extension, Meta,
and Resource. Each type's definition includes a reference
to FHIR documentation.

The value[x] field of UsageContext is not fully defined, as some of the
possible types for the value are not yet defined.

A UsageContext's value may be a Range.
Information about the Range FHIR type can be found at:
http://hl7.org/fhir/R4/datatypes.html#Range
The validateRange method will throw an error if given a Range that
violates a constraint.