Flag rules should not mark unset flags as false
To be clearer about what was intentionally set in the FSH, we should leave unset flags as `undefined`.  This makes a difference because if a base class has a flag set to true, not setting it in a child class doesn't mean to change it to false -- it means to leave it as-is.