Reorganize types, add FHIRdateTime
Primitive types, data types, and metadata types are combined into one file for each group.
Each of these files is in turn loaded by the index in the fhirtypes directory.
The FHIRdateTime represents a primitive FHIR type. Tests are included to ensure conformance
with the regular expression provided in the FHIR specification.