Clean up of comments, imports, and tests
Combine import statements for types that originate in the same file.
Change "FHIRdateTime" to "FHIRDateTime", because that better adheres to capitalization convention.
The validation method's name is similarly changed to "validateFHIRDateTime".
Remove extraneous comments.
When variables are used only once in test cases, use the values directly rather than creating the variables in a higher scope.