Add various FHIR types (#4)
* Add various FHIR types

New types are: UsageContext, Range, Identifier, Period,
ContactDetail, ContactPoint, Narrative, Extension, Meta,
and Resource. Each type's definition includes a reference
to FHIR documentation.

The value[x] field of UsageContext is not fully defined, as some of the
possible types for the value are not yet defined.

A UsageContext's value may be a Range.
Information about the Range FHIR type can be found at:
http://hl7.org/fhir/R4/datatypes.html#Range
The validateRange method will throw an error if given a Range that
violates a constraint.

* Reorganize types, add FHIRdateTime

Primitive types, data types, and metadata types are combined into one file for each group.
Each of these files is in turn loaded by the index in the fhirtypes directory.
The FHIRdateTime represents a primitive FHIR type. Tests are included to ensure conformance
with the regular expression provided in the FHIR specification.

* Fix import of Meta type

* Remove unused variables from Period test

Also, remove unnecessary type annotation from dateTime primitive regex.

* Clean up of comments, imports, and tests

Combine import statements for types that originate in the same file.
Change "FHIRdateTime" to "FHIRDateTime", because that better adheres to capitalization convention.
The validation method's name is similarly changed to "validateFHIRDateTime".
Remove extraneous comments.
When variables are used only once in test cases, use the values directly rather than creating the variables in a higher scope.

* Add Reference type. Consolidate and add documentation links to types.