Throw errors when setting isModifier or mustSupport from true to false
These errors cover some of the rules about setting flags. The team is still working
to make sure we have a complete understanding of the rules about flags, so there
may be some modifications in the future.

The test case for not modifying a profile when mustSupport goes from true to false
is skipped for now, as it requires the creation of a new structure definition file.