Apply FlagRules when exporting a structure definition (#14)
* Apply FlagRule when exporting

Current implementation only updates the flags. There are conditions which affect
when these flags can be changed, but checks for those conditions are not yet
implemented.

* Throw errors when setting isModifier or mustSupport from true to false

These errors cover some of the rules about setting flags. The team is still working
to make sure we have a complete understanding of the rules about flags, so there
may be some modifications in the future.

The test case for not modifying a profile when mustSupport goes from true to false
is skipped for now, as it requires the creation of a new structure definition file.

* Add test for preventing mustSupport changing from true to false

Nothing in the FHIR spec has mustSupport set to true, so a custom resource is used in the test.

* "undefined" values do not affect Element flags

When applyFlags receives an undefined argument, the flag on the
element will not be changed.
Update a test case to use a FHIR profile instead of a custom
structure definition.

* Fix merge issue