Fixes needed to support npm publish
In order to support npm publish we need to ensure that we do publish the right things (compiled js, json, and *.d.ts definition files) and we don't publish the wrong things (original ts files, map files, antlr jar, etc).  We also need to use a name that isn't already in use at npm (sushi *is* already in use).

To support this, the following changes were made:
- package.json: changed "name" to "fsh-sushi"
- package.json: added a "prepare" script to run "npm build" before packaging (to ensure compiled js is up to date)
- package.json: added a "prepublishOnly" script to run "npm run check" before publishing (to ensure we don't publish code w/ errors)
- package.json: added "files" to whitelist what we want published
- package.json: added "types" to point to the main *.d.ts file (currently empty, but that's ok)
- tsconfig.json: set "declaration" to true to generate .d.ts files in dist
- importText.ts: fixed bad relative paths in import statements