Add file source information to FshEntity
Storing both the file and the text location is necessary for useful logging. The SourceInfo type contains these
attributes. Checking the file is added to a test case in FSHImporter Profile test suite. Functions to set a
FshEntity's file and location are added in order to improve readability of FSHImporter and associated
test cases.
A side effect of the updates to FshEntity is that it is no longer possible to create implicit FshCode objects,
as was typically the case when throwing CodeAlreadyFixedError in ElementDefinition.