Add source tracking for FSH entities (#30)
* Add start and end attributes to Profile

When a Profile is created by FSH import, store the line and column
for where the Profile's definition starts and ends.

* Add start and end attributes to Extension

* Create function for obtaining start and stop values from a context

* Set FshEntity parent, get location from TerminalNode, add location to tests

FshEntity is added as a parent across classes in fshtype, allowing them to track a location.
TerminalNode instances keep their location data in a slightly different way than other ParserRuleContext instances.

* Add file source information to FshEntity

Storing both the file and the text location is necessary for useful logging. The SourceInfo type contains these
attributes. Checking the file is added to a test case in FSHImporter Profile test suite. Functions to set a
FshEntity's file and location are added in order to improve readability of FSHImporter and associated
test cases.
A side effect of the updates to FshEntity is that it is no longer possible to create implicit FshCode objects,
as was typically the case when throwing CodeAlreadyFixedError in ElementDefinition.