Allow alias types for instances, enforce metadata rules
The InstanceOf metadata field may contain an alias, so it is resolved while parsing.
The InstanceOf metadata field is required on Instances, so throw an error if it is missing.
Metadata fields on Instances should only be declared once, so throw an error if they are declared more than once.