Include FSH source information in log statements (#36)
* Defined WithSource interface for errors

Errors that are thrown due to incorrect FSH should implement the WithSource interface,
which will allow the logger to display information regarding the file name and text location
that contains the erroneous FSH text.

* Test for source information in ProfileExporter

When exporting a profile whose parent isn't found, and that profile has
source information, expect the source information to be present in the
logged error message.

* Check logging in StructureDefinitionExporter tests

StructureDefinitionExporter catches and logs numerous errors when
exporting. Therefore, the tests do not try to directly catch the errors,
but instead checks that expected source information is logged.

* Log source location for unsupported rule

Currently, obeys rules are not supported. Once all rule types are
supported, the test for this message will have to be re-examined.

* Use source information in extension exporter

Extension source information logging is similar to profile source
information logging.