Import instances when parsing FSH (#33)
* Basic Instance import

Instances can be defined in FSH and imported into SUSHI. Currently, the values specified
for the instance are not actually processed and stored in the Instance object.

* Parse FixedValueRules for Instances

The rules on an Instance can only be FixedValueRule, and not any other type of rule.

* Allow alias types for instances, enforce metadata rules

The InstanceOf metadata field may contain an alias, so it is resolved while parsing.
The InstanceOf metadata field is required on Instances, so throw an error if it is missing.
Metadata fields on Instances should only be declared once, so throw an error if they are declared more than once.

* Add duplicate metadata checks to Profile and Extension

Each metadata field should only be declared once on a given Profile or Extension.
Because the constructor for Extension assigns values to id and parent, an additional variable is used
to track whether those metadata fields have been declared.

* Do not stop processing on duplicate metadata

* Catch and log RequiredMetadataError for Instance

Processing should continue when an Instance without a type is
encountered. The error is caught, no instance is added to the document,
and processing continues.