Define grammar for ValueSet
A ValueSet is an entity that contains one or mode codes. An
Element that contains a code may be constrained to only take codes
from a specific ValueSet. The members of a ValueSet may be specific
codes, or they may be defined based on descent from other codes.

This update to the grammar does not include any changes to the
importer, and as a result, there are no associated test cases.

Additionally, the definition of lexer rule CODE is changed, as the
STRING? is not used.