Compose ValueSet with inclusions and exclusions
The members of a ValueSet are specified by including and excluding
terms, filters, and ValueSets. A filter must have a system and may
have a property name upon which to operate. This design will not
accomodate the full expressive power of the FHIR specifications
for value set composition, but it will allow for all the rules
for value sets to be fulfilled.