Add ValueSet class (#47)
* Add ValueSet class

The FSH ValueSet class is used to represent value sets. Value sets
may contain specific codes. Value sets may also specify that they
include or exclude codes that are descendents of a specified code.

* Fix spelling of "descendants"

* Add ValueSetTerm class for members of ValueSet

New terms defined as part of a ValueSet may optionally contain a
definition. Definitions are not included when codes are used in
other contexts, so FshCode is extended to include this field.

* Compose ValueSet with inclusions and exclusions

The members of a ValueSet are specified by including and excluding
terms, filters, and ValueSets. A filter must have a system and may
have a property name upon which to operate. This design will not
accomodate the full expressive power of the FHIR specifications
for value set composition, but it will allow for all the rules
for value sets to be fulfilled.

* Use ValueSetComponent to compose a ValueSet

The possible entries in a ValueSet's inclusion and exclusion lists
are represented as instances of the new ValueSetComponent class.
The different composable rules are represented using three new
types: ValueSetConcept, ValueSetFilter, and ValueSetReference.
This more closely matches the FHIR specifications than before.