ValueSet grammar component recognition
The ValueSet grammar is restructured to provide more accurate
and detailed parsing. A ValueSet contains one or more components,
each of which is a term, reference, or filter. The grammar for
matching a filter is the most complex, as there are many legal
forms a filter may take. It is the responsibility of the import
functions to ensure that a grammatically correct filter is also
semantically correct. Some comments in the grammar specification
file are included to help clarify the rules for filters.