Support more components in ValueSet grammar
The FHIR specification for what can be included or excluded in a
value set should now be fully supported by the FSH grammar. Previous
versions did not fully support the specified cardinalities for
value set components.

New metadata fields are available when defining a value set.

These changes are based on feedback and discussion during the review
of the contents of this branch.

The relevant FHIR specification is located at:
http://hl7.org/fhir/valueset-definitions.html#ValueSet.compose.include