Add ValueSet to grammar (#54)
* Define grammar for ValueSet

A ValueSet is an entity that contains one or mode codes. An
Element that contains a code may be constrained to only take codes
from a specific ValueSet. The members of a ValueSet may be specific
codes, or they may be defined based on descent from other codes.

This update to the grammar does not include any changes to the
importer, and as a result, there are no associated test cases.

Additionally, the definition of lexer rule CODE is changed, as the
STRING? is not used.

* ValueSet grammar component recognition

The ValueSet grammar is restructured to provide more accurate
and detailed parsing. A ValueSet contains one or more components,
each of which is a term, reference, or filter. The grammar for
matching a filter is the most complex, as there are many legal
forms a filter may take. It is the responsibility of the import
functions to ensure that a grammatically correct filter is also
semantically correct. Some comments in the grammar specification
file are included to help clarify the rules for filters.

* Allow ValueSet filter with only Operator and Value

A ValueSet filter of "descendant-of AAA#123" is allowable.
Therefore, the grammar is updated to make both System and
Property optional. Checking the validity of the filter will be the
responsibility of the importer.

* Support more components in ValueSet grammar

The FHIR specification for what can be included or excluded in a
value set should now be fully supported by the FSH grammar. Previous
versions did not fully support the specified cardinalities for
value set components.

New metadata fields are available when defining a value set.

These changes are based on feedback and discussion during the review
of the contents of this branch.

The relevant FHIR specification is located at:
http://hl7.org/fhir/valueset-definitions.html#ValueSet.compose.include

* Remove flags from regex in ValueSet grammar

Alas, although flags would be useful when specifying a regular
expression in a value set filter, the FHIR specifications do not
appear to support this. The flags are thusly removed.

* Separate concepts with comma in value set grammar

When listing multiple concepts in a single value set component,
they should be separated by a comma followed by at least one whitespace
character. A from clause is always required when listing multiple
concepts, so the following component is no longer valid:
* ABC#123, ABC#234
Instead, it should be specified as one of the following:
With a from clause,
* #123, #234 from system ABC
or as separate components,
* ABC#123
* ABC#234

* Allow spaces in concepts in ValueSet grammar

A specified concept string is permitted to contain non-consecutive
whitespace characters. Users are required to surround such strings
in double-quote characters in order to define such strings. The
quoted concept string can contain double-quote characters by escaping
them with a backslash character, similarly to the definition for the
STRING lexer rule.

* Fix double-quote escaping in strings

The grammar pertaining to strings enclosed by double-quotes has been
updated. They should now correctly allow strings to contain
double-quote and backslash characters by preceding either with an
additional backslash.

* Unescape double-quote and backslash in importer

When the importer handles a STRING token or quoted concept, it will
remove the surrounding quotation marks, then unescape all backslash
and double-quote characters in the string.