CodeSystem grammar (#61)
* CodeSystem grammar

The grammar definition now includes CodeSystems. A CodeSystem has
a few metadata fields, and contains a list of codes. Existing rules
provide most of the definition of this new entity.

* CodeSystem concepts can include a definition

When defining concepts as part of a code system, a definition string
can be included. If present, it will appear after the description
string.

Co-authored-by: Chris Moesel <cmoesel@users.noreply.github.com>