Remove duplicate definitions from Package
There is currently a bug in how we do exports that causes some duplicates in the  Package.  More specifically, if a FSH Extension is resolved while exporting a Profile,  then the resolved Extension will be put in the Profiles array.  The reverse is also true.  We need to determine how best to fix that bug, but in the meantime, we will just remove the duplicates when we return a Package in FHIRExporter.