Add tests for ValueSet import filters
Each filter operator has an associated positive test case. Because
the exists filter expects a boolean value, the keywords for true
and false are added as options in the parser rule for the filter
value. The string array type is removed from filter values since
FHIR does not support it, which means there is no value in parsing
a string of comma-separated values into an array.
Test cases for errors that may occur when importing a value set
with filters are not yet added.