Check ValueSet filter value types
Each filter operator expects to receive a value of a certain type.
Log an error and skip the filter if a value of a different type
is received. Regular expression values are now stored using the
RegExp type instead of as strings, as this makes it easier to confirm
that the appropriate type is being used.