Import ValueSet (#68)
* Basic import for ValueSet entities

The FSH importer will parse value set definitions. Only one test
case so far, for a very simple value set.

* Process value set metadata when importing

The available metadata fields are id, title, and description.
At this point, there are three very similar looking blocks for the
three types of metadata, so it may be high time to reconsider the
structuring of metadata processing.

* Initial support for ValueSetFilterComponent import

Basic instances of ValueSetFilterComponent without any actual
filters can be imported. The codes can be included from a single
system, one or more value sets, or both.

Additionally, the FHIR specification permits value set definitions
without any include/exclude statements, so the grammar now allows
this as well.

* Import value set with basic filter

Very basic filter importing is supported. Property and operator
must be members of VsProperty and VsOperator, respectively.
Value has been simplified somewhat, but will later be enhanced
as test coverage for filters is increased.

* Add tests for ValueSet import filters

Each filter operator has an associated positive test case. Because
the exists filter expects a boolean value, the keywords for true
and false are added as options in the parser rule for the filter
value. The string array type is removed from filter values since
FHIR does not support it, which means there is no value in parsing
a string of comma-separated values into an array.
Test cases for errors that may occur when importing a value set
with filters are not yet added.

* Check ValueSet filter value types

Each filter operator expects to receive a value of a certain type.
Log an error and skip the filter if a value of a different type
is received. Regular expression values are now stored using the
RegExp type instead of as strings, as this makes it easier to confirm
that the appropriate type is being used.

* Add sourceInfo checks to some ValueSet import tests

* Test for invalid filter property or operator

* ValueSet import test coverage

Test cases are added for:
- duplicate metadata
- multiple codes in a comma-separated list
- missing filter value (only allowed for "exists" operator)

Test description for "descendant-of" revised to clarify that it
is an alternate allowed spelling for the "descendent-of" operator.

* Add test for ValueSet component with multiple filters