ValueSet validation, additional export tests
Validation of FHIR-defined rules for value sets occurs both during
import and export. Most rules are checked during import, as they
pertain to the components from which a value set's included and
excluded sets are composed. The rule that a value set with a compose
element must include at least one component is enforced during
export.