Export formatting, merge value set components
Exported value sets include the 'resourceType' field. The value of
this field is always 'ValueSet'.
Value sets are sorted by name when being added to the IG definition.
When multiple concept components in a value set have the same
inclusion and are from the same system and value sets, the concepts
in those components are merged. This reduces the number of components
in the value set without affecting the definition.