Export value set (#77)
* Basic ValueSet export

Value sets are included in the export process. The previous FSH
ValueSet class is renamed to FshValueSet, and the new FHIR ValueSet
class is named ValueSet. Validation is not yet implemented. Only
minimal tests have been added.

* ValueSet validation, additional export tests

Validation of FHIR-defined rules for value sets occurs both during
import and export. Most rules are checked during import, as they
pertain to the components from which a value set's included and
excluded sets are composed. The rule that a value set with a compose
element must include at least one component is enforced during
export.

* Additional tests for value set export

* Resolve aliases for value set sources

Code systems and value sets used in a value set's "from" clause
may be aliased.

* ValueSet filter value shows code, but not system

* Export ValueSet in IG, format/naming corrections

ValueSets are included in the exported json files when SUSHI is run
from command line. No extra steps are necessary to prepare the
ValueSet object, so the toJSON method is minimal.

Names of various types used by the ValueSet class are changed to be
more consistent with the naming conventions used for other FHIR
types.

ValueSetComposeError's message is revised for clarity.

ValueSetExporter's setCompose has been cleaned up, and now creates a
compose object without an exclude field if there are no excluded
elements. The unused FHIRDefs are removed from the constructor.

Tests for the ValueSetExporter now compare the exporter's result
using an object equality test rather than testing individual fields.
Test values for systems and value sets are now URLs, which more
closely resemble expected data.

* Display is optional when composing ValueSet

When a ValueSet composition includes concepts, the code field is
required, but the display field is not. If no display value is
available, the field should not be present on the exported object.

* Create type alias for ValueSet filter value

* Export formatting, merge value set components

Exported value sets include the 'resourceType' field. The value of
this field is always 'ValueSet'.
Value sets are sorted by name when being added to the IG definition.
When multiple concept components in a value set have the same
inclusion and are from the same system and value sets, the concepts
in those components are merged. This reduces the number of components
in the value set without affecting the definition.

* Allow any string in ValueSetFilter property

The ValueSet filter is allowed to be any string.  In addition, replace the `code` property in filter tests with `concept` -- as that is the property SNOMED-CT uses.

* Minor stylistic tweak

Co-authored-by: Chris Moesel <cmoesel@users.noreply.github.com>