Add source information to log messages, make best effort with instance rules (#119)
* Source info in Instance and StructureDefinition export

Source information is present in some export messages that did not
have that information previously. A side effect is that instances
with incorrect rules will still have their correct rules applied.

When a fixed value rule's path is not found on an instance, the
error message now contains source information. When an extension
is sliced, and that slice does not have a url (either specified
or by reference), the error message now contains source information.
Tests are added for both of these error messages.

* Source info for circular dependency scenarios

When StructureDefinitionExporter emits a circular dependency
warning, include source information.
When MasterFisher emits a circular dependency error, include source
information if the circular dependency originates in a FSH
structure definition.