Show message on invalid FHIR id or name
The id and name elements in code systems, structure definitions,
value sets, and instances are checked against the invariants from
the FHIR specification. Both invariants are regular expressions.
When an id or name is not valid, it is still used in the output,
but a message indicates to the user that it is invalid.

The validation of these elements occurs when exporting to FHIR.
Although it would be possible to check some of this earlier on
during FSH import, the validation step is based on specifications
from FHIR. Therefore, I think it makes the most sense to enforce
these rules at the time of creating the FHIR resources.