Update logic to copy dependency info into ImplementationGuide JSON
If the version is "current" or "dev", don't attempt to match on the ig.version, because it will be a real version number (not "current" or "dev").  Also, do not encode the version number in the "uri", just use the base canonical uri and put the requested version in "version".