Refactor id and name validation
FHIR classes with id or name attributes now use the HasName and
HasId mixins. CodeSystem, StructureDefinition, and ValueSet all
use the mixins in the same way. InstanceDefinition has an id, but
no name, and as such only mixes in the HasId class. Additionally,
because the id on an InstanceDefinition may be updated by setting
a fixed value, an additional validation method is present on the
HasId class in order to validate the existing property.

The error messages for invalid name and id values now contain
additional information to clarify what counts as valid for those
fields.

Un-reordered the attributes in InstanceDefinition, which were
accidentally reordered in a previous commit.