Show message on invalid FHIR id or name (#142)
* Show message on invalid FHIR id or name

The id and name elements in code systems, structure definitions,
value sets, and instances are checked against the invariants from
the FHIR specification. Both invariants are regular expressions.
When an id or name is not valid, it is still used in the output,
but a message indicates to the user that it is invalid.

The validation of these elements occurs when exporting to FHIR.
Although it would be possible to check some of this earlier on
during FSH import, the validation step is based on specifications
from FHIR. Therefore, I think it makes the most sense to enforce
these rules at the time of creating the FHIR resources.

* Refactor id and name validation

FHIR classes with id or name attributes now use the HasName and
HasId mixins. CodeSystem, StructureDefinition, and ValueSet all
use the mixins in the same way. InstanceDefinition has an id, but
no name, and as such only mixes in the HasId class. Additionally,
because the id on an InstanceDefinition may be updated by setting
a fixed value, an additional validation method is present on the
HasId class in order to validate the existing property.

The error messages for invalid name and id values now contain
additional information to clarify what counts as valid for those
fields.

Un-reordered the attributes in InstanceDefinition, which were
accidentally reordered in a previous commit.

* Avoid duplicate property declaration in FHIR types

FHIR types that receive the id and name properties from mixins do
not need to declare those properties directly.

ElementDefinition's id property turned back to string in order to
correctly match its specification as being FHIR primitive type
string, not FHIR primitive type id as other id properties are.