Info messages for task start and package load
The tasks that have info messages when they finish now also have
info messages when they start. This should aid in troubleshooting.

Package loading now includes messages pertaining to checking the
local cache and downloading files. As package loading is an
asynchronous operation, the exact ordering of these messages is not
guaranteed to be consistent. This information will help users know
when SUSHI is able to load local files and when it must spend time
downloading packages.