Escape # in coding system identifier (#208)
* Allow escaped # in system identifiers in codings

The # may be present in a system identifier. This is particularly
likely when the identifier is a URL. However, since # is also used
to separate the system and the code, it is necessary to use a
backslash character to escape it and prevent it from being used as
the separator. Because backslash is now an escape character, it is
also necessary to escape literal backslash characters.

* Support edge cases for coding system during import

The separating # character may be preceded by escaped backslashes.
The coding system may be only escaped backslashes.