Sort pages in IG content (#226)
* Sort pagecontent with numeric prefixes

Files in IG pagecontent can have an order specified by starting the
file name with a number followed by an underscore. Files with numeric
prefixes are added to the page list in ascending order, followed by
files without numeric prefixes. If two files have the same numeric
prefix, the file names are compared to determine the order.

* Prevent file name collisions in IG content

When a numeric prefix is removed from an IG content file, it is
possible that a file name collision could occur. To prevent this,
use the original name of the file, including the prefix.

* Define local type for PageMetadata