Rename index page title to Home
We previously set the index page title to the IG title, but in some cases the IG title is fairly long and more specific than the home page name needs to be.  Now we just set the page title to "Home" - which should work for any IG.

When we revisit IG config, this will likely change again to a user-configurable value.