Import Invariant and ObeysRule (#238)
* Import Invariant

FSH importer can now add invariants. Invariants contain several
metadata fields, but no additional information.

* Import ObeysRule

An ObeysRule on a Profile may contain a path to one of the Profile's
elements and will contain the names of one or more Invariants.

* Add test for Extension that contains an ObeysRule

* Log error for invalid Invariant severity

Severity codes are bound to ValueSet ConstraintSeverity based on
the specification of ElementDefinition. Only "error" and "warning"
are allowed.

* Refactor visitSeverity and visitCode in FSHImporter

Both of these parser rules include a CODE lexeme, so the logic for
handling that lexeme is in a new parseCodeLexeme method.

* Invariant requires Description and Severity

Errors are logged when an Invariant is missing a description or a
severity code.