Detect rules that ineffective due to type constraint
When a type constraint is applied to a choice element, some of the
existing choices may no longer be possible. If the same profile has
rules that eliminate a choice and apply rules to that choice, the
rules that apply to the eliminated choice have no real effect. In
these cases, show the user an error message, as it is likely to be
evidence of incorrect specification.

Note that if rules are applied to a choice in a parent profile, no
message is shown, as the rule in the parent profile still has an
effect.