Show a message when created choices are made obsolete with type constraints (#253)
* Detect rules that ineffective due to type constraint

When a type constraint is applied to a choice element, some of the
existing choices may no longer be possible. If the same profile has
rules that eliminate a choice and apply rules to that choice, the
rules that apply to the eliminated choice have no real effect. In
these cases, show the user an error message, as it is likely to be
evidence of incorrect specification.

Note that if rules are applied to a choice in a parent profile, no
message is shown, as the rule in the parent profile still has an
effect.

* Handle constraint application to neighbor slices

When a sliced element contains a choice element, a profile may
define multiple slices on that element that each have different
types for the contained choice element. In this case, no error has
occurred, and no message should be logged.

* Use parent function to help find obsolete choices

Comments added to this branch's test cases containing the FSH that
would generate the rules specified.