Change strategy for constraints and slices interfering
The previous implementation of always logging an error when
fixing a value on an element with a slicing was too broad. The
new strategy will involve actually checking if fixing a value or
applying other constraints will actually cause a problem. The
previous tests involving fixed values and slices are skipped until
the relevant implementations are completed.

Additionally, it is now an error to constrain the cardinality of
an element that exists inside slices if the new cardinality on the
root element is narrower than the cardinality on a connected element
contained within a slice.