Change handling of ValueSetRule on slices
A slice can have a completely different value set bound to it than
the list element being sliced. But, if the slice and the list are
bound to the same value set, then the binding on the slice must be
at least as strong as the binding on the list.