Allow standards status to change
A new standards status flag is allowed to replace an existing
standards status flag.

The mustSupport, summary, and modifier flags can be enabled, but not
disabled. False is now ignored just like undefined is when applying
flags to an element.

Remove error classes for replacing a standards status and for
disabling mustSupport and modifier flags.