Log error when fixing value on sliced element (#259)
* Log error when fixing value on sliced element

When an element has slicing defined on it, log an error when a value
is fixed on the element's root path. Fixing a value on the root path
in this situation is occasionally legal, but almost never correct.

* Change strategy for constraints and slices interfering

The previous implementation of always logging an error when
fixing a value on an element with a slicing was too broad. The
new strategy will involve actually checking if fixing a value or
applying other constraints will actually cause a problem. The
previous tests involving fixed values and slices are skipped until
the relevant implementations are completed.

Additionally, it is now an error to constrain the cardinality of
an element that exists inside slices if the new cardinality on the
root element is narrower than the cardinality on a connected element
contained within a slice.

* Add FlagRule and slice tests

Some tests for FlagRule and other rules are still in progress.

* Update slice when root element has FlagRule applied

The function findConnectedElements is updated to also return elements
that are directly slices of the element that calls the function.
The FlagRule test case added in this commit demonstrates an example
of this scenario.

* Test invalid FlagRule on sliced list

* Test case for invalid FlagRule on list's child

* Handle ValueSetRules on sliced list elements

When possible, ValueSetRules applied to sliced list elements should
be applied to their defined slices. In the case where the defined
slice already has a stronger binding, do not apply the ValueSetRule.
No error is logged in this case.

* Apply ObeysRule to slices and their children

* Handle type constraints on choice elements in slices

It is permissible to eliminate a choice via type constraint, but
not to eliminate all types on a slice.

* Change handling of ValueSetRule on slices

A slice can have a completely different value set bound to it than
the list element being sliced. But, if the slice and the list are
bound to the same value set, then the binding on the slice must be
at least as strong as the binding on the list.

* Allow null on connected element cardinality

When checking the cardinality of connected elements, if the min or
max on the connected element is null, then no error should occur.

Update an OnlyRule test case for removing all types to check the
logger for an appropriate error.

* Better path split for findConnectedElements

Using splitOnPathPeriods will avoid problems when the name of a
slice contains periods.

Two more tests added for binding value sets to children of list
elements and children of slices.