Non-breaking space character counts as whitespace (#342)
* Non-breaking space character counts as whitespace

The character with decimal value 160 represents a non-breaking
space, and is now grouped in with other whitespace characters in
the grammar. This helps in situations where FSH is copied from a
web page that adds some non-breaking space characters as part of
the formatting.

* Add NONWS and NONWS_STR fragments

These new fragments help simplify a few other grammar rules.