Allow version in concept system for ValueSet (#353)
* Allow version in concept system for ValueSet

When declaring a concept component for a ValueSet, the system uri
may be followed by a pipe and a version string. This is aligned with
other scenarios that use a system uri.

* Set version attribute in value set components

When exporting a value set, the include the version attribute in the
objects within the include or exclude arrays.

I really wanted to do this with some acrobatic one-liner, but I was
not acrobatic enough.